refactor: extract shared error inspector initializer
