+++ lib/chef/formatters/error_inspectors/base_error_inspector.rb
@@ -1,0 +1,33 @@
+#--
+# Copyright:: Copyright (c) 2009-2026 Progress Software Corporation and/or its subsidiaries or affiliates. All Rights Reserved.
+# License:: Apache License, Version 2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+class Chef
+  module Formatters
+    module ErrorInspectors
+      module BaseErrorInspector
+        attr_reader :exception
+        attr_reader :node_name
+        attr_reader :config
+
+        def initialize(node_name, exception, config)
+          @node_name = node_name
+          @exception = exception
+          @config = config
+        end
+      end
+    end
+  end
+end
