+++ lib/chef/formatters/error_inspectors/node_load_error_inspector.rb
@@ -17,6 +17,7 @@
 #
 
 require_relative "api_error_formatting"
+require_relative "base_error_inspector"
 require "chef-utils/dist" unless defined?(ChefUtils::Dist)
 
 class Chef
@@ -28,16 +29,7 @@ class Chef
       class NodeLoadErrorInspector
 
         include APIErrorFormatting
-
-        attr_reader :exception
-        attr_reader :node_name
-        attr_reader :config
-
-        def initialize(node_name, exception, config)
-          @node_name = node_name
-          @exception = exception
-          @config = config
-        end
+        include BaseErrorInspector
 
         def add_explanation(error_description)
           case exception
